12th commit
Added an enemy ship, enemy laser and collision detection. The collision
detection is still a bit buggy. I didn't use the original enemy ship i
found as i kept running into problems/errors with it involving its
collision detection, code for it that effected the lighting of the game
and it disappearing in play mode.